// src/@types/config.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 *
 * Generated via https://borischerny.com/json-schema-to-typescript-browser/
 */

/**
 * Name for this WLED instance.
 */
export type WLEDName = string;
/**
 * IP address for this WLED instance.
 */
export type WLEDIPAddress = string;
export type Wleds = WLEDConfiguration[];

export interface WLEDPluginConfiguration {
  wleds?: Wleds;
  [k: string]: unknown;
}
export interface WLEDConfiguration {
  name: WLEDName;
  ip: WLEDIPAddress;
  [k: string]: unknown;
}